+++ src/insert/spacefinder/article.ts
@@ -17,13 +17,22 @@ import {
 	getCurrentBreakpoint,
 	getCurrentTweakpoint,
 } from 'lib/detect/detect-breakpoint';
-import { waitForAdvert } from '../../lib/dfp/wait-for-advert';
+import { isInVariantSynchronous } from '../../experiments/ab';
+import { deeplyReadRightColumn } from '../../experiments/tests/deeply-read-right-column';
 import fastdom from '../../utils/fastdom-promise';
 import { computeStickyHeights, insertHeightStyles } from '../sticky-inlines';
 import { initCarrot } from './carrot-traffic-driver';
 import { isInHighValueSection } from './utils';
 
 type SlotName = Parameters<typeof createAdSlot>[0];
+
+/**
+ * As estimation of the height of the most viewed island.
+ * This appears from desktop breakpoints on the right-hand side.
+ * Knowing the height of the element is useful when
+ * calculating where to place ads in the right column.
+ */
+const MOST_VIEWED_HEIGHT = 600;
 
 const articleBodySelector = '.article-body-commercial-selector';
 
@@ -34,13 +43,13 @@ const hasImages = !!window.guardian.config.page.lightboxImages?.images.length;
 const hasShowcaseMainElement =
 	window.guardian.config.page.hasShowcaseMainElement;
 
-const isInMegaTestControl =
-	window.guardian.config.tests?.commercialMegaTestControl === 'control';
-
-const increaseAdDensity = isInHighValueSection && !isInMegaTestControl;
+const isInDeeplyReadMostViewedVariant = isInVariantSynchronous(
+	deeplyReadRightColumn,
+	'deeply-read-and-most-viewed',
+);
 
 const minDistanceBetweenRightRailAds = 500;
-const minDistanceBetweenInlineAds = increaseAdDensity ? 500 : 750;
+const minDistanceBetweenInlineAds = isInHighValueSection ? 500 : 750;
 
 /**
  * Rules to avoid inserting ads in the right rail too close to each other
@@ -160,8 +169,8 @@ const addDesktopInline1 = (fillSlot: FillAdSlot): Promise<boolean> => {
 		opponentSelectorRules: {
 			// don't place ads right after a heading
 			':scope > h2, [data-spacefinder-role="nested"] > h2': {
-				minAboveSlot: increaseAdDensity ? 150 : 5,
-				minBelowSlot: increaseAdDensity ? 0 : 190,
+				minAboveSlot: isInHighValueSection ? 150 : 5,
+				minBelowSlot: isInHighValueSection ? 0 : 190,
 			},
 			[`.${adSlotContainerClass}`]: {
 				minAboveSlot: 500,
@@ -176,7 +185,7 @@ const addDesktopInline1 = (fillSlot: FillAdSlot): Promise<boolean> => {
 				minBelowSlot: 600,
 			},
 			'figure.element--supporting': {
-				minAboveSlot: 500,
+				minAboveSlot: 100,
 				minBelowSlot: 0,
 			},
 		},
@@ -214,11 +223,12 @@ const addDesktopRightRailAds = (fillSlot: FillAdSlot): Promise<boolean> => {
 
 	/**
 	 * In special cases, inline2 can overlap the "Most viewed" island, so
-	 * we need to make an adjustment to move the inline2 further down the page.
+	 * we need to make an adjustment to move the inline2 further down the page
 	 */
-	if (isPaidContent) {
-		minAbove += 600;
+	if (isInDeeplyReadMostViewedVariant || isPaidContent) {
+		minAbove += MOST_VIEWED_HEIGHT;
 	}
+
 	// Some old articles don't have a main image, which means the first paragraph is much higher
 	if (!hasImages) {
 		minAbove += 600;
@@ -310,64 +320,66 @@ const addDesktopRightRailAds = (fillSlot: FillAdSlot): Promise<boolean> => {
 	});
 };
 
-const addMobileInlineAds = (fillSlot: FillAdSlot): Promise<boolean> => {
-	const minDistanceFromArticleTop = !isInMegaTestControl ? 100 : 200;
+const mobileMinDistanceFromArticleTop = 200;
 
-	/**
-	 * These 2 sets of rules are for the changes to "ranked" articles as part of the mega test
-	 */
-	const oldRules: SpacefinderRules = {
-		bodySelector: articleBodySelector,
-		candidateSelector: ':scope > p',
-		minAbove: 200,
-		minBelow: 200,
-		opponentSelectorRules: {
-			':scope > h2': {
-				minAboveSlot: 100,
-				minBelowSlot: 250,
-			},
-			...inlineAdSlotContainerRules,
-			[`:scope > :not(p):not(h2):not(.${adSlotContainerClass}):not(#sign-in-gate)`]:
-				{
-					minAboveSlot: 35,
-					minBelowSlot: 200,
-				},
+const mobileIgnoreList = `:not(p):not(h2):not(hr):not(.${adSlotContainerClass}):not(#sign-in-gate):not([data-spacefinder-type$="NumberedTitleBlockElement"])`;
+
+const mobileOpponentSelectorRules = {
+	// don't place ads right after a heading
+	':scope > h2, [data-spacefinder-role="nested"] > h2, :scope > [data-spacefinder-type$="NumberedTitleBlockElement"]':
+		{
+			minAboveSlot: 100,
+			minBelowSlot: 0,
 		},
-		filter: (candidate, lastWinner) => {
-			if (!lastWinner) {
-				return true;
-			}
-			const distanceBetweenAds = candidate.top - lastWinner.top;
-			return distanceBetweenAds >= minDistanceBetweenInlineAds;
+	...inlineAdSlotContainerRules,
+	// this is a catch-all for elements that are not covered by the above rules, these will generally be things like videos, embeds and atoms. minBelowSlot is higher to push ads a bit further down after these elements
+	[`:scope > ${mobileIgnoreList}, [data-spacefinder-role="nested"] > ${mobileIgnoreList}`]:
+		{
+			minAboveSlot: 35,
+			minBelowSlot: 200,
+			// Usually we don't want an ad right before videos, embeds and atoms etc. so that we don't break up related content too much. But if we have a heading above, anything above the heading won't be related to the current content, so we can place an ad there.
+			bypassMinBelow:
+				'h2,[data-spacefinder-type$="NumberedTitleBlockElement"]',
 		},
-	};
+};
 
-	const ignoreList = `:not(p):not(h2):not(hr):not(.${adSlotContainerClass}):not(#sign-in-gate):not([data-spacefinder-type$="NumberedTitleBlockElement"])`;
-
-	const newRules: SpacefinderRules = {
+const addMobileTopAboveNav = (fillSlot: FillAdSlot): Promise<boolean> => {
+	const rules: SpacefinderRules = {
 		bodySelector: articleBodySelector,
 		candidateSelector:
 			':scope > p, :scope > h2, :scope > [data-spacefinder-type$="NumberedTitleBlockElement"], [data-spacefinder-role="nested"] > p',
-		minAbove: minDistanceFromArticleTop,
+		minAbove: mobileMinDistanceFromArticleTop,
 		minBelow: 200,
-		opponentSelectorRules: {
-			// don't place ads right after a heading
-			':scope > h2, [data-spacefinder-role="nested"] > h2, :scope > [data-spacefinder-type$="NumberedTitleBlockElement"]':
-				{
-					minAboveSlot: 100,
-					minBelowSlot: 0,
-				},
-			...inlineAdSlotContainerRules,
-			// this is a catch-all for elements that are not covered by the above rules, these will generally be things like videos, embeds and atoms. minBelowSlot is higher to push ads a bit further down after these elements
-			[`:scope > ${ignoreList}, [data-spacefinder-role="nested"] > ${ignoreList}`]:
-				{
-					minAboveSlot: 35,
-					minBelowSlot: 200,
-					// Usually we don't want an ad right before videos, embeds and atoms etc. so that we don't break up related content too much. But if we have a heading above, anything above the heading won't be related to the current content, so we can place an ad there.
-					bypassMinBelow:
-						'h2,[data-spacefinder-type$="NumberedTitleBlockElement"]',
-				},
-		},
+		opponentSelectorRules: mobileOpponentSelectorRules,
+	};
+
+	const insertAds: SpacefinderWriter = async (paras) => {
+		const slots = paras.slice(0, 1).map(async (para) => {
+			const name = 'top-above-nav';
+			const type = 'top-above-nav';
+			const slot = await insertSlotAtPara(para, name, type, 'inline');
+			return fillSlot(name, slot);
+		});
+		await Promise.all(slots);
+	};
+
+	return spaceFiller.fillSpace(rules, insertAds, {
+		waitForImages: true,
+		waitForInteractives: true,
+		pass: 'mobile-top-above-nav',
+	});
+};
+
+const addMobileSubsequentInlineAds = (
+	fillSlot: FillAdSlot,
+): Promise<boolean> => {
+	const rules: SpacefinderRules = {
+		bodySelector: articleBodySelector,
+		candidateSelector:
+			':scope > p, :scope > h2, :scope > [data-spacefinder-type$="NumberedTitleBlockElement"], [data-spacefinder-role="nested"] > p',
+		minAbove: mobileMinDistanceFromArticleTop,
+		minBelow: 200,
+		opponentSelectorRules: mobileOpponentSelectorRules,
 		/**
 		 * Filter out any candidates that are too close to the last winner
 		 * see https://github.com/guardian/commercial/tree/main/docs/spacefinder#avoiding-other-winning-candidates
@@ -382,12 +394,10 @@ const addMobileInlineAds = (fillSlot: FillAdSlot): Promise<boolean> => {
 		},
 	};
 
-	const rules = isInMegaTestControl ? oldRules : newRules;
-
 	const insertAds: SpacefinderWriter = async (paras) => {
 		const slots = paras.map(async (para, i) => {
-			const name = i === 0 ? 'top-above-nav' : `inline${i}`;
-			const type = i === 0 ? 'top-above-nav' : 'inline';
+			const name = `inline${i + 1}`;
+			const type = 'inline';
 			const slot = await insertSlotAtPara(para, name, type, 'inline');
 			return fillSlot(
 				name,
@@ -406,7 +416,7 @@ const addMobileInlineAds = (fillSlot: FillAdSlot): Promise<boolean> => {
 	return spaceFiller.fillSpace(rules, insertAds, {
 		waitForImages: true,
 		waitForInteractives: true,
-		pass: 'mobile-inlines',
+		pass: 'mobile-subsequent-inlines',
 	});
 };
 
@@ -415,21 +425,26 @@ const addMobileInlineAds = (fillSlot: FillAdSlot): Promise<boolean> => {
  * @param fillSlot A function to call that will fill the slot when each ad slot has been inserted,
  * these could be google display ads or opt opt consentless ads.
  */
-const addInlineAds = (fillSlot: FillAdSlot): Promise<boolean> => {
+const addFirstInlineAd = (fillSlot: FillAdSlot): Promise<boolean> => {
 	const isMobile = getCurrentBreakpoint() === 'mobile';
 	if (isMobile) {
-		return addMobileInlineAds(fillSlot);
+		return addMobileTopAboveNav(fillSlot);
 	}
 
 	if (isPaidContent) {
-		return addDesktopRightRailAds(fillSlot);
+		return Promise.resolve(false);
 	}
 
-	// Add the rest of the inline ad slots after a position for inline1 has been found.
-	// We don't wan't inline1 and inline2 targeting the same paragraph.
-	return addDesktopInline1(fillSlot).then(() =>
-		addDesktopRightRailAds(fillSlot),
-	);
+	return addDesktopInline1(fillSlot);
+};
+
+const addSubsequentInlineAds = (fillSlot: FillAdSlot): Promise<boolean> => {
+	const isMobile = getCurrentBreakpoint() === 'mobile';
+	if (isMobile) {
+		return addMobileSubsequentInlineAds(fillSlot);
+	}
+
+	return addDesktopRightRailAds(fillSlot);
 };
 
 const attemptToAddInlineMerchAd = (
@@ -493,22 +508,22 @@ const attemptToAddInlineMerchAd = (
  * Init all the article body adverts, including `im` and `carrot`
  * @param fillAdSlot a function to fill the ad slots
  */
-const init = async (fillAdSlot: FillAdSlot): Promise<boolean> => {
+const init = async (fillAdSlot: FillAdSlot): Promise<void> => {
 	if (!commercialFeatures.articleBodyAdverts) {
-		return Promise.resolve(false);
+		return Promise.resolve();
 	}
 
-	await addInlineAds(fillAdSlot);
+	// We add the first inline ad before finding a place for an im so that the inline1 doesn't get pushed down
+	await addFirstInlineAd(fillAdSlot);
 
-	const im = window.guardian.config.page.hasInlineMerchandise
-		? attemptToAddInlineMerchAd(fillAdSlot)
-		: Promise.resolve(false);
-	const inlineMerchAdded = await im;
-	if (inlineMerchAdded) await waitForAdvert('dfp-ad--im');
+	if (window.guardian.config.page.hasInlineMerchandise) {
+		await attemptToAddInlineMerchAd(fillAdSlot);
+	}
+
+	// Add the other inline slots after the im, so that there is space left for the im ad
+	await addSubsequentInlineAds(fillAdSlot);
 
 	await initCarrot();
-
-	return im;
 };
 
-export { init, addInlineAds, type FillAdSlot };
+export { init, addFirstInlineAd, addSubsequentInlineAds, type FillAdSlot };