+++ src/insert/spacefinder/spacefinder-debug-tools.ts
@@ -1,5 +1,4 @@
 import { log } from '@guardian/libs';
-import { getUrlVars } from 'utils/url';
 import type {
 	SpacefinderExclusions,
 	SpacefinderItem,
@@ -7,8 +6,6 @@ import type {
 	SpacefinderRules,
 } from './spacefinder';
 import logo from './spacefinder-logo.svg';
-
-const enableDebug = !!getUrlVars().sfdebug;
 
 const isCurrentPass = (pass: SpacefinderPass) => {
 	const sfdebugPass = document.querySelector<HTMLInputElement>(
@@ -58,6 +55,10 @@ const exclusionTypes = {
 		colour: colours.blue,
 		reason: 'Too close to other element',
 	},
+	overlaps: {
+		colour: colours.blue,
+		reason: 'Overlaps other element',
+	},
 } as const;
 
 const isExclusionType = (type: string): type is keyof typeof exclusionTypes =>
@@ -72,7 +73,9 @@ const addOverlay = (element: HTMLElement, text: string) => {
 
 const addHoverListener = (
 	candidate: HTMLElement,
-	tooClose: Exclude<SpacefinderItem['meta'], undefined>['tooClose'],
+	tooClose: Exclude<SpacefinderItem['meta'], undefined>[
+		| 'tooClose'
+		| 'overlaps'],
 	pass: SpacefinderPass,
 ) => {
 	tooClose.forEach((opponent) => {
@@ -100,10 +103,12 @@ const addHoverListener = (
 			}
 
 			opponent.element.classList.add('blocking-element');
-			addOverlay(
-				opponent.element,
-				`${opponent.actual}px/${opponent.required}px`,
-			);
+			if (opponent.actual && opponent.required) {
+				addOverlay(
+					opponent.element,
+					`${opponent.actual}px/${opponent.required}px`,
+				);
+			}
 		});
 
 		candidate.addEventListener('mouseleave', () => {
@@ -138,6 +143,9 @@ const annotateExclusions = (
 				element.setAttribute(`data-sfdebug-${pass}`, 'isStartAt');
 			} else if (type) {
 				element.setAttribute(`data-sfdebug-${pass}`, key);
+			} else if (meta && meta.overlaps.length > 0) {
+				element.setAttribute(`data-sfdebug-${pass}`, 'overlaps');
+				addHoverListener(element, meta.overlaps, pass);
 			} else if (meta && meta.tooClose.length > 0) {
 				element.setAttribute(`data-sfdebug-${pass}`, 'tooClose');
 				addHoverListener(element, meta.tooClose, pass);
@@ -407,8 +415,6 @@ const init = (
 	rules: SpacefinderRules,
 	pass: SpacefinderPass,
 ): void => {
-	if (!enableDebug) return;
-
 	addDebugPanel();
 	annotate(exclusions, winners, rules, pass);
 	addPassToDebugPanel(pass);