+++ src/lib/header-bidding/slot-config.ts
@@ -188,7 +188,9 @@ const getSlots = (): HeaderBiddingSizeMapping => {
 			],
 		},
 		'mobile-sticky': {
-			mobile: shouldIncludeMobileSticky() ? [adSizes.mobilesticky] : [],
+			mobile: shouldIncludeMobileSticky()
+				? [adSizes.mobilesticky, createAdSize(300, 50)]
+				: [],
 		},
 		'crossword-banner': {
 			desktop: isCrossword ? [adSizes.leaderboard] : [],
@@ -196,6 +198,14 @@ const getSlots = (): HeaderBiddingSizeMapping => {
 		},
 		'crossword-banner-mobile': {
 			mobile: [adSizes.mobilesticky],
+		},
+		'football-right': {
+			desktop: [
+				adSizes.empty,
+				adSizes.mpu,
+				adSizes.skyscraper,
+				adSizes.halfPage,
+			],
 		},
 		merchandising: {
 			mobile: [adSizes.mpu],