+++ webpack.config.prod.js
@@ -16,11 +16,13 @@ const gitCommitSHA = () => {
 	}
 };
 
+const prefix = process.env.BUNDLE_PREFIX ?? '[chunkhash]/';
+
 module.exports = webpackMerge.smart(config, {
 	mode: 'production',
 	output: {
-		filename: `[chunkhash]/graun.standalone.commercial.js`,
-		chunkFilename: `[chunkhash]/graun.[name].commercial.js`,
+		filename: `${prefix}graun.standalone.commercial.js`,
+		chunkFilename: `${prefix}graun.[name].commercial.js`,
 		path: path.join(__dirname, 'dist', 'bundle', 'prod'),
 		clean: true,
 	},